+++ babel.config.js
@@ -1,0 +1,14 @@
+// babel.config.js
+module.exports = {
+  presets: [
+    [
+      "@babel/preset-env",
+      {
+        targets: {
+          node: "current"
+        }
+      }
+    ],
+    "@babel/preset-typescript"
+  ]
+};
