+++ jest.config.js
@@ -1,0 +1,13 @@
+// For a detailed explanation regarding each configuration property, visit:
+// https://jestjs.io/docs/en/configuration.html
+
+module.exports = {
+  // The test environment that will be used for testing
+  testEnvironment: "jsdom",
+
+  // A map from regular expressions to paths to transformers
+  transform: {
+    "^.+\\.tsx?$": "ts-jest"
+  },
+  testMatch: ["**/*.test.ts"]
+};
