+++ src/index.test.ts
@@ -1,0 +1,3 @@
+describe("Just leave me alone Jest", () => {
+  it("it is fine, go away", () => expect(1).toEqual(1));
+});
