--- test/index.ts
@@ -1,7 +1,0 @@
-import { arbitraryQuery } from "../output/output";
-
-describe("Test fast-check generation", () => {
-  it("Runs the tests at all", () => {
-    expect(true).toBeTruthy();
-  });
-});
